+++ index.html
@@ -8,7 +8,6 @@
   <title>Кексобукинг</title>
 </head>
 <body>
-  <script src="/js/main.js"></script>
   <main>
     <div class="promo">
       <h1 class="promo__title visually-hidden">Keksobooking. Кексы по соседству</h1>
@@ -230,6 +229,6 @@
       <button type="button" class="error__button">Попробовать снова</button>
     </div>
   </template>
-
+  <script src="/js/main.js" type="module"></script>
 </body>
 </html>
